+++ Assets/XCharts/Runtime/Internal/CoordinateChart_DrawHeatmap.cs
@@ -131,8 +131,8 @@ namespace XCharts
             var rangeMax = m_VisualMap.rangeMax;
             var color = m_ThemeInfo.GetColor(serie.index);
             var borderWidth = serie.itemStyle.show ? serie.itemStyle.borderWidth : 0;
-            var borderColor = serie.itemStyle.opacity > 0 ? serie.itemStyle.borderColor : Color.clear;
-            borderColor.a *= serie.itemStyle.opacity;
+            var borderColor = serie.itemStyle.opacity > 0 ? serie.itemStyle.borderColor : ChartConst.clearColor32;
+            borderColor.a = (byte)(borderColor.a * serie.itemStyle.opacity);
             serie.dataPoints.Clear();
             serie.animation.InitProgress(1, 0, xCount);
             var animationIndex = serie.animation.GetCurrIndex();
@@ -151,7 +151,7 @@ namespace XCharts
                         serie.dataPoints.Add(Vector3.zero);
                         continue;
                     }
-                    var value = serieData.GetCurrData(dimension, dataChangeDuration, yAxis.inverse);
+                    var value = serieData.GetCurrData(dimension, dataChangeDuration, yAxis.inverse, yAxis.runtimeMinValue, yAxis.runtimeMaxValue);
                     if (serieData.IsDataChanged()) dataChanging = true;
                     var pos = new Vector3(zeroX + (i + 0.5f) * xWidth, zeroY + (j + 0.5f) * yWidth);
                     serie.dataPoints.Add(pos);
@@ -181,7 +181,7 @@ namespace XCharts
                     if (m_VisualMap.hoverLink && emphasis && serie.emphasis.show && serie.emphasis.itemStyle.borderWidth > 0)
                     {
                         var emphasisBorderWidth = serie.emphasis.itemStyle.borderWidth;
-                        var emphasisBorderColor = serie.emphasis.itemStyle.opacity > 0 ? serie.emphasis.itemStyle.borderColor : Color.clear;
+                        var emphasisBorderColor = serie.emphasis.itemStyle.opacity > 0 ? serie.emphasis.itemStyle.borderColor : ChartConst.clearColor32;
                         ChartDrawer.DrawBorder(vh, pos, rectWid, rectHig, emphasisBorderWidth, emphasisBorderColor);
                     }
                 }